move-lightnode-file-dir
